+++ src/modpack/mp_create.rs
@@ -45,7 +45,8 @@ pub fn collect_mp_create_args(args: &MPCreateArgs) -> Result<ModInfo, RustiqueEr
 // When a modpack is created, the mods by default, will be moved to their own folder in modpacks/installed/mynewpack
 // If those mods were created from a symlink,
 
-pub async fn mp_create(mod_dir: impl PathRef + Copy, mod_pack: &mut ModInfo, save_location: Option<impl PathRef>, copy_mods: bool, ignore_modpacks: bool) -> Result<(PathBuf, PathBuf), RustiqueError> {
+#[allow(clippy::fn_params_excessive_bools)]
+pub async fn mp_create(mod_dir: impl PathRef + Copy, mod_pack: &mut ModInfo, save_location: Option<impl PathRef>, copy_mods: bool, ignore_modpacks: bool, include_configs: bool, copy_configs: bool) -> Result<(PathBuf, PathBuf), RustiqueError> {
     
     let config = get_config().write().await;
     let modpack_dir = config.modpacks.modpack_dir.clone();