+++ src/modpack/mp_enable.rs
@@ -1,13 +1,8 @@
 
-
-// Enabling a modpack is a bit of work. First it mvs the existing default mod directory to the rustique backup location,
-// then create a symlink to the modpack. Only one modpack can be enabled at a time
-
 use std::path::Path;
 use comfy_table::{Attribute, Color};
 use comfy_table::presets::UTF8_HORIZONTAL_ONLY;
-use crate::aliases::FileName;
-use crate::commands::arg_structs::modpack_args::MPEnableArgs;
+use crate::aliases::{FileName, ModID};
 use crate::config::config_manager::get_config;
 use crate::information_utils::{display_table, notice, CellData};
 use crate::modpack::symlink_manager::SymlinkManager;
@@ -15,10 +10,23 @@ use crate::rustique_errors::RustiqueError;
 use crate::traits::ref_ext::PathRef;
 use crate::utils::extract_all_mods_metadata;
 
-pub async fn mp_enable(args: MPEnableArgs, mod_dir: impl PathRef) -> Result<String, RustiqueError> {
-    let config = get_config().read().await;
+#[cfg(windows)]
+use is_elevated::is_elevated;
+
+#[cfg(windows)]
+use std::process::exit;
+
+pub async fn mp_enable(mpk_id: ModID, mod_dir: impl PathRef, force: bool) -> Result<String, RustiqueError> {
     
-    let mod_pack_install_dir = Path::new(&config.modpacks.modpack_dir).join("installed").join(&args.mpk_id);
+   
+    #[cfg(windows)]
+    if !is_elevated() {
+        notice("In order to enable modpacks, Rustique uses symlinks which require admin permissions on Windows. Please run Rustique with admins right and try again.", Some(Color::Red), vec![Attribute::Bold]);
+        exit(1);
+    }
+    
+    let config = get_config().read().await;
+    let mod_pack_install_dir = Path::new(&config.modpacks.modpack_dir).join("installed").join(&mpk_id);
     
     if !mod_pack_install_dir.exists() {
         return Err(RustiqueError::SimpleError("Modpack {} doesn't exist. Run 'Rustique modpack list' or 'Rustique modpack local list' to view your installed modpacks.".into()));
@@ -28,19 +36,19 @@ pub async fn mp_enable(args: MPEnableArgs, mod_dir: impl PathRef) -> Result<Stri
     // if so, notify the user and tell them to either disable the current one OR use modpack enable -f to force the use and warn about using multiple
     
     // Is it already enabled?
-    if config.modpacks.enabled.contains(&args.mpk_id) {
-        notice(format!("Modpack: [{}] is already enabled. Did you mean to disable it?", &args.mpk_id), Some(Color::Yellow), vec![Attribute::Bold]);
+    if config.modpacks.enabled.contains(&mpk_id) {
+        notice(format!("Modpack: [{}] is already enabled. Did you mean to disable it?", &mpk_id), Some(Color::Yellow), vec![Attribute::Bold]);
         return Err(RustiqueError::SimpleError("Modpack already enabled".into()));
     }
     
     // Is it even installed??
-    if !config.modpacks.disabled.contains(&args.mpk_id) {
-        notice(format!("Modpack: [{}] is not installed! Use [Rustique modpack install {}] to install it first.", &args.mpk_id, &args.mpk_id), Some(Color::Yellow), vec![Attribute::Bold]);
+    if !config.modpacks.disabled.contains(&mpk_id) {
+        notice(format!("Modpack: [{}] is not installed! Use [Rustique modpack install {}] to install it first.", &mpk_id, &mpk_id), Some(Color::Yellow), vec![Attribute::Bold]);
         return Err(RustiqueError::SimpleError("Modpack needs to be installed first".into()))
     }
     
     // Is anything else enabled?
-    if !config.modpacks.enabled.is_empty() && !&args.force {
+    if !config.modpacks.enabled.is_empty() && !force {
         
         display_table(vec![
             (CellData::new("You already have the following modpack(s) enabled: ".into(), Some(Color::Yellow), vec![], None),
@@ -71,5 +79,5 @@ pub async fn mp_enable(args: MPEnableArgs, mod_dir: impl PathRef) -> Result<Stri
     }
     
     
-    Ok(args.mpk_id)
+    Ok(mpk_id)
 }